tasks 1-5 using ViewSets + division by application
