+++ drf_project/authors/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class DrfAppConfig(AppConfig):
+class AuthorsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'drf_app'
+    name = 'authors'
