+++ drf_project/authors/models.py
@@ -14,11 +14,3 @@ class Author(models.Model):
     """
     def get_id_of_books(self):
         return list(self.author_of_books.values_list('id', flat=True))
-
-
-class Book(models.Model):
-    title = models.CharField(max_length=255)
-    author = models.ForeignKey(Author, related_name='author_of_books', on_delete=models.CASCADE, default=None)
-
-    def __str__(self):
-        return self.title
