+++ drf_project/authors/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Author
+
+
+class AuthorSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Author
+        fields = '__all__'
