+++ drf_project/authors/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path, include
+from .views import AuthorsViewSet
+
+from rest_framework import routers
+router = routers.SimpleRouter()
+router.register(r'', AuthorsViewSet)
+
+
+urlpatterns = [
+    path('', include(router.urls)),
+]
