+++ drf_project/books/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class BooksConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'books'
