+++ drf_project/books/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Book
+
+
+class BookSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Book
+        fields = '__all__'
