--- drf_project/drf_app/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from .models import Author, Book
-
-admin.site.register(Author)
-admin.site.register(Book)
