--- drf_project/drf_app/forms.py
@@ -1,1 +1,0 @@
-from django import forms
