--- drf_project/drf_app/serializers.py
@@ -1,15 +1,0 @@
-from rest_framework import serializers
-from .models import Author, Book
-
-
-class AuthorSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Author
-        fields = '__all__'
-
-
-class BookSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Book
-        fields = '__all__'
-
