--- drf_project/drf_app/urls.py
@@ -1,28 +1,0 @@
-"""
-URL configuration for drf_project project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
-from django.urls import path
-from .views import books_list_view, books_detail_view, authors_list_view, authors_detail_view, author_books_view
-
-
-urlpatterns = [
-    path('books/', books_list_view, name='books_list'),
-    path('books/<int:pk>', books_detail_view, name='books_detail'),
-    path('authors/', authors_list_view, name='authors_list'),
-    path('authors/<int:pk>/', authors_detail_view, name='authors_detail'),
-    path('author_books/<int:author_id>/', author_books_view, name='author_books'),
-]
