+++ drf_project/drf_project/settings.py
@@ -36,7 +36,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'drf_app',
+    'authors',
+    'books',
     'rest_framework',
 ]
